Create a library for screen handling
